review commit + rename HyperspaceConf
